LPS-102245 Move WeavingClassLoader and URLWeavingAdapter to portal-test; they are only used by tests
